spes: parts/thermal: Drop intent filter and profile reset
After switching to TaskStackListener, the behavior of resetting
profile to default on SCREEN_OFF intent is problematic because
there is no mechanism to set it back once screen turns on again.
The reason is that onTaskStackChanged() is not trigerred when
you simply turn off and turn back on the screen when an app is
in foreground. As a result, if a foreground app has non-default
thermal profile, after a screen-off and screen-on cycle, the
profile is changed to default and does not come back.

We have two options:
1. Bring back SCREEN_ON intent filter again, remember the profile
   at SCREEN_OFF intent, reset the profile to default, and restore
   the profile after SCREEN_ON.
2. Do not reset the profile to default at all.

I prefer the second option because resetting to default profile
provides no meaningful functionality. Sure, the thermal profile may
remain at a non-default entry when screen is off, but that only
means the thermal-engine/mi_thermald may have a different set of
thermal mitigation thresholds and actions. Apps are not supposed
to be running long and intensive work when screen is off anyway.
If the foreground app is moved to background or killed during the
screen-off period, onTaskStackChanged() will set the profile
accordingly.

Signed-off-by: Chenyang Zhong <zhongcy95@gmail.com>